PaymentLogo handling of web-payment
This fix prevents warnings for an invalid payment type value being given
to the PaymentLogo component.

You can see the warning on the plan page provided web-payment is a valid
payment type for your user / geo etc

PaymentLogo is used to display a list of logo on the plan page.

The logo list is decided by what payments are avaialble for the user so
web-payment should be covered.

Todo:
[ ] Test the logo / css - can't see it with this commit alone
[ ] Confirm alt text is correct (just guessed, not super obvious what it
should be, check commit that introduced it?)